cleanup code some more
